refactor: перенос функции update_admin_info в sub_hl к другим аналогичным функциям
